Refactor: Move ACI logic into Azure.Deployments.Extensibility.Providers.ThirdParty library
