--- .devcontainer/Dockerfile
@@ -1,1 +1,0 @@
-FROM mcr.microsoft.com/dotnet/sdk:6.0
