+++ Azure.Deployments.Extensibility.sln
@@ -25,6 +25,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DevHost", "DevHost", "{D067
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Deployments.Extensibility.DevHost", "src\Azure.Deployments.Extensibility.DevHost\Azure.Deployments.Extensibility.DevHost.csproj", "{33BEB45A-04F9-4824-AFE1-2B20B439EA67}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{7D35F9AA-A555-4ACD-BC97-0F49EE7E6DF4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Deployments.Extensibility.Providers.ThirdParty", "src\Azure.Deployments.Extensibility.Providers.ThirdParty\Azure.Deployments.Extensibility.Providers.ThirdParty.csproj", "{C9AA69F5-2811-4F73-A559-1560150623B8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -59,6 +63,10 @@ Global
 		{33BEB45A-04F9-4824-AFE1-2B20B439EA67}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{33BEB45A-04F9-4824-AFE1-2B20B439EA67}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{33BEB45A-04F9-4824-AFE1-2B20B439EA67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C9AA69F5-2811-4F73-A559-1560150623B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C9AA69F5-2811-4F73-A559-1560150623B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C9AA69F5-2811-4F73-A559-1560150623B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C9AA69F5-2811-4F73-A559-1560150623B8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -72,6 +80,7 @@ Global
 		{CEA4E336-491B-4D1A-A578-776E380EF463} = {13E01E57-50E3-42B7-A0E2-5EA0709CB27B}
 		{CE6AF926-A2F5-41C3-B6CD-84D76AAF50CA} = {13E01E57-50E3-42B7-A0E2-5EA0709CB27B}
 		{33BEB45A-04F9-4824-AFE1-2B20B439EA67} = {D067770F-49EE-4374-B878-C624450A7329}
+		{C9AA69F5-2811-4F73-A559-1560150623B8} = {7D35F9AA-A555-4ACD-BC97-0F49EE7E6DF4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E3073660-079C-482C-ABB4-E970087314E7}
