+++ src/Azure.Deployments.Extensibility.Core/IExtensibilityProvidersRegistry.cs
@@ -6,7 +6,5 @@ namespace Azure.Deployments.Extensibility.Core
     public interface IExtensibilityProviderRegistry
     {
         IExtensibilityProvider? TryGetExtensibilityProvider(string providerName);
-
-        ExtensibilityProviderContainerRegistry? TryGetExtensibilityProviderContainerRegistry(string providerName);
     }
 }
